Add flags to set speed unit to mps or kph

+++ include/hmi_message_publisher/planning_general_info_msg.h
@@ -19,20 +19,20 @@ static const google::protobuf::EnumDescriptor *TrajectoryType_descriptor = drive
 class PlanningGeneralInfo {
     public:
         PlanningGeneralInfo();
-        PlanningGeneralInfo(const bool& allow_lane_change, const bool& left_lane_change_cmd, const bool& right_lane_change_cmd, const double& set_speed, const double& speed_limit);
+        PlanningGeneralInfo(const bool& allow_lane_change, const bool& left_lane_change_cmd, const bool& right_lane_change_cmd, const int& set_speed, const int& speed_limit);
         PlanningGeneralInfo(const VectorDataType& data_vector);   
        
         void setAllowLaneChange(const bool& allow_lane_change);
         void setLeftLaneChangeCmd(const bool& left_lane_change_cmd);
         void setRightLaneChangeCmd(const bool& right_lane_change_cmd);
-        void setSetSpeed(const double& set_speed);
-        void setSpeedLimit(const double& speed_limit);
+        void setSetSpeed(const int& set_speed);
+        void setSpeedLimit(const int& speed_limit);
         
         bool getAllowLaneChange();
         bool getLeftLaneChangeCmd();
         bool getRightLaneChangeCmd();
-        double getSetSpeed();
-        double getSpeedLimit();
+        int getSetSpeed();
+        int getSpeedLimit();
         
         VectorDataType getVectorData() const { return _data_vec; }
