+++ src/planning_general_info_msg.cpp
@@ -13,7 +13,7 @@ namespace hmi_message{
         setSpeedLimit(0);
     }
 
-    PlanningGeneralInfo::PlanningGeneralInfo(const bool& allow_lane_change, const bool& left_lane_change_cmd, const bool& right_lane_change_cmd, const double& set_speed, const double& speed_limit){
+    PlanningGeneralInfo::PlanningGeneralInfo(const bool& allow_lane_change, const bool& left_lane_change_cmd, const bool& right_lane_change_cmd, const int& set_speed, const int& speed_limit){
         _data_vec.resize(PlanningGeneralInfo_data_size);
 
         setAllowLaneChange(allow_lane_change);
@@ -52,19 +52,19 @@ namespace hmi_message{
         return PlanningGeneralInfo_right_lane_change_cmd.recoverValueFromDataVector(_data_vec);
     }
 
-    void PlanningGeneralInfo::setSetSpeed(const double& set_speed) {
+    void PlanningGeneralInfo::setSetSpeed(const int& set_speed) {
         PlanningGeneralInfo_set_speed.writeValueToDataVector(_data_vec, set_speed);
     }
 
-    double PlanningGeneralInfo::getSetSpeed() {
+    int PlanningGeneralInfo::getSetSpeed() {
         return PlanningGeneralInfo_set_speed.recoverValueFromDataVector(_data_vec);
     }
 
-    void PlanningGeneralInfo::setSpeedLimit(const double& speed_limit) {
+    void PlanningGeneralInfo::setSpeedLimit(const int& speed_limit) {
         PlanningGeneralInfo_speed_limit.writeValueToDataVector(_data_vec, speed_limit);
     }
 
-    double PlanningGeneralInfo::getSpeedLimit() {
+    int PlanningGeneralInfo::getSpeedLimit() {
         return PlanningGeneralInfo_speed_limit.recoverValueFromDataVector(_data_vec);
     }
